+++ src/Symfony/Component/VarDumper/Caster/SocketCaster.php
@@ -13,36 +13,21 @@ namespace Symfony\Component\VarDumper\Caster;
 
 use Symfony\Component\VarDumper\Cloner\Stub;
 
-/**
- * @author Nicolas Grekas <p@tchwork.com>
- * @author Alexandre Daubois <alex.daubois@gmail.com>
- */
 final class SocketCaster
 {
-    public static function castSocket(\Socket $h, array $a, Stub $stub, bool $isNested): array
+    public static function castSocket(\Socket $socket, array $a, Stub $stub, bool $isNested): array
     {
-        socket_getsockname($h, $addr, $port);
-        $info = stream_get_meta_data(socket_export_stream($h));
+        socket_getsockname($socket, $addr, $port);
+        $info = stream_get_meta_data(socket_export_stream($socket));
 
         $a += [
             Caster::PREFIX_VIRTUAL.'address' => $addr,
             Caster::PREFIX_VIRTUAL.'port' => $port,
-            Caster::PREFIX_VIRTUAL.'info' => new EnumStub(array_intersect_key(
-                $info,
-                [
-                    'timed_out' => new ConstStub($info['timed_out'] ? 'true' : 'false'),
-                    'blocked' => new ConstStub($info['blocked'] ? 'true' : 'false'),
-                    'eof' => new ConstStub($info['eof'] ? 'true' : 'false'),
-                    'unread_bytes' => new ScalarStub($info['unread_bytes']),
-                    'stream_type' => new ConstStub($info['stream_type']),
-                    'wrapper_type' => new ConstStub($info['wrapper_type'] ?? ''),
-                    'wrapper_data' => new ConstStub($info['wrapper_data'] ?? ''),
-                    'mode' => new ConstStub($info['mode']),
-                    'seekable' => new ConstStub($info['seekable'] ? 'true' : 'false'),
-                    'uri' => new ConstStub($info['uri'] ?? ''),
-                ]
-            )),
         ];
+
+        foreach ($info as $key => $val) {
+            $a[Caster::PREFIX_VIRTUAL.$key] = $val;
+        }
 
         return $a;
     }