+++ src/Symfony/Component/VarDumper/Caster/SqliteCaster.php
@@ -16,14 +16,17 @@ use Symfony\Component\VarDumper\Cloner\Stub;
 /**
  * @author Alexandre Daubois <alex.daubois@gmail.com>
  */
-class SqliteCaster
+final class SqliteCaster
 {
-    public static function castSqlite3Result(\SQLite3Result $c, array $a, Stub $stub, bool $isNested): array
+    public static function castSqlite3Result(\SQLite3Result $result, array $a, Stub $stub, bool $isNested): array
     {
         $a += [
-            Caster::PREFIX_VIRTUAL.'numColumns' => $c->numColumns(),
-            Caster::PREFIX_VIRTUAL.'result' => $c->fetchArray(\SQLITE3_ASSOC),
+            Caster::PREFIX_VIRTUAL.'numColumns' => $result->numColumns(),
         ];
+
+        for ($i = 0; $i < $result->numColumns(); ++$i) {
+            $a[Caster::PREFIX_VIRTUAL.'columnName'][$i] = $result->columnName($i);
+        }
 
         return $a;
     }